NP-13634 Show journal fields for ReaderOpinion and FeatureArticle
